test,refactor: move grouping tests
